December 12, 2015 + navbar + Footer + Contact modal
The navigation bar is now fixed to the top of the page.
Removed Useful Links menu.
Removed Useful Links information from the About page. The Useful Links
menu has been replaced with an Up arrow which jumps to the top of the
page.
Removed the Survey banner.
Removed the Up arrow in the footer.