April 10, 2016 + Ajaxed footer
The terminal can now be accessed using the tilde key on the keyboard.
The terminal is now loaded in the automatically in the background.
Added more commands to the terminal.

The footer is now ajaxed.
Added new easter egg to namePrompt function.
Added Keybase account to the contact modal.
Removed eye test image previously found in the hide ads modal.
Tweaked various modals.