April 15, 2016 + New select menu styles + Footer id
Added new commands to the Terminal.
The snowstorm command allows you to change settings with the snow Easter
egg.
The water command allows you to change settings with the water Easter
egg.

Implemented the bootstrap-select plugin to add extra styles to the
select menus.
Not too sure how this'll work on mobile. But hey, it looks great on
desktop.
Added custom highlighting colors.