October 7, 2017
The Search and namePrompt function assets are now ajaxed when needed.
This reduces page load times.

Added an option to Toggle the Terminal on page.
This is shown only if the Load Terminal assets option is checked.

Added the ability to determine if an error in loading major ajax calls occurs.
Most ajax requests are now only done once to prevent further requests from the server.
This will reduce server load.