October 20, 2017
Migrated to localForage instead of cookies for storing Settings and Achievements.
This allows for better compatibility between browsers and reduces server load.
Added the ability to clear all local storage in the Stats modal.

Removed the local site warning alerts as they are no longer needed.
Renamed the alerts to have a prefix of alert.

Replaced the cookies command with the storage command in the Terminal.
It functions mostly the same but there are some differences.
Make sure to view help for more information.

Updated Bootstrap to beta 2.
Updated jQuery Terminal to v1.9.0.