October 27, 2017
Added "Requests" section to the Stats modal.
This displays the number of unique IP requests to the server (so not necessarily page visits).

Added new functions to the Terminal.
IP, Agent, and Display.
View help for more information.
Added the ability to interpret JavaScript.

The Privacy and Terms banners have been replaced with static messages on the top right of the page.
Cleaned up some JavaScript.