December 8, 2017
Added a new Achievement called "Unlimited POWER".
The Achievement Method is now a coulumn in the table rather than a popover.

Added new commands to the Terminal.
A noteworthy one is the history command which allows you to view and clear command history.
Added a proper white theme to the Terminal.
Cleaned up some CSS by using inheritence for the menus.