April 28, 2018
Search on mobile now uses a modal for better usability.
Added new commands and major command tweaks to the Terminal.
The requests command returns unique and total requests to the site.
The libraries command returns the current version of JavaScript libraries used.
The media command will now only function if you're on the homepage and audio is present.
The modal command now prevents the user from opening multiple modals and closing hidden ones.
The anchor command has new syntax and the ability to list all the IDs on the page.
The snowstorm command has more help information.

Updated localForage to version 1.7.1.
Updated Keypress to version 5.0.10.
Updated Konami-JS to version 1.6.0.