May 25, 2018
PNotify is now being used for alerts/notifications.
This comes with better mobile support and smoother animations compared to bootstrapGrowl.
The alert command on the Terminal now interfaces with this.

The "Cookie" banner has been replaced with the "Legal" banner.
This now the only place to accept the Terms & Conditions and Privacy Policy.

Updated Bootstrap to version 4.1.1.
Removed the "Silent Mode" from Settings.
The footer has been tweaked.