October 12, 2018
Content and Date Jump menus are now fixed when the page is scrolled far enough.
Scrolling to anchors should no longer clip content.
Removed achievement from items in the achievement local storage database (as it was redundant).
This change will cause issues with achievement progress thus resetting progress is recommended.
The legal banner is now an alert.

Minor improvements to the Terminal.
Added bg command to fix background color of the Terminal not being applied.
The close command now allows the animation to be played.